backend/requirements.txt
flask
flask-cors
python-dotenv
PyMuPDF
openai
google-genai
requests
deep-translator
fpdf2
python-docx
ollama
weasyprint
anthropic
mistralai
moviepy